TO RESTAURANT
feat: add Saved Order Notes management page and functionality

- Implemented SavedOrderNotesPage for managing saved order notes.
- Added API integration for creating, updating, and deleting saved order notes.
- Included form validation and user feedback using toast notifications.
- Created a table to display saved order notes with edit and delete options.

feat: add Tables Management page and functionality

- Implemented TablesManagementPage for managing restaurant tables.
- Added API integration for creating, updating, and deleting restaurant tables.
- Included form validation and user feedback using toast notifications.
- Created a table to display restaurant tables with edit and delete options.

feat: add Tables Grid page for table selection

- Implemented TablesGridPage to display available and occupied tables.
- Added navigation to manage tables and handle table selection for orders.
- Included user feedback for table status and actions.

feat: define Recipe types for recipe management

- Created recipe.types.ts to define Recipe and RecipeIngredient interfaces.
- Added types for creating and updating recipes.

feat: define Restaurant types for table and order note management

- Created restaurant.types.ts to define RestaurantTable and SavedOrderNote interfaces.
- Added types for creating and updating restaurant tables and saved order notes.

feat: implement order status flow utility functions

- Created orderStatusFlow.ts to manage order status transitions and actions.
- Added functions to determine next operational status and valid actions for orders.

+++ backend/KasserPro.Application/DTOs/Inventory/BranchInventoryDto.cs
@@ -9,8 +9,8 @@ public class BranchInventoryDto
     public string ProductName { get; set; } = string.Empty;
     public string? ProductSku { get; set; }
     public string? ProductBarcode { get; set; }
-    public int Quantity { get; set; }
-    public int? BatchAvailableQuantity { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal? BatchAvailableQuantity { get; set; }
     public int ReorderLevel { get; set; }
     public bool IsLowStock { get; set; }
     public DateTime LastUpdatedAt { get; set; }
@@ -26,6 +26,6 @@ public class BranchInventorySummaryDto
     public int ProductId { get; set; }
     public string ProductName { get; set; } = string.Empty;
     public string? ProductSku { get; set; }
-    public int TotalQuantity { get; set; }
+    public decimal TotalQuantity { get; set; }
     public List<BranchInventoryDto> BranchInventories { get; set; } = new();
 }
