+++ backend/KasserPro.Application/DTOs/Products/AdjustStockRequest.cs
@@ -2,14 +2,14 @@ namespace KasserPro.Application.DTOs.Products;
 
 public class AdjustStockRequest
 {
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string Reason { get; set; } = string.Empty;
     public string AdjustmentType { get; set; } = "Adjustment";
 }
 
 public class StockAdjustResultDto
 {
-    public int NewBalance { get; set; }
-    public int PreviousBalance { get; set; }
-    public int Change { get; set; }
+    public decimal NewBalance { get; set; }
+    public decimal PreviousBalance { get; set; }
+    public decimal Change { get; set; }
 }
