+++ backend/KasserPro.Application/DTOs/Reports/InventoryReportDto.cs
@@ -8,7 +8,7 @@ public class BranchInventoryReportDto
     public int BranchId { get; set; }
     public string BranchName { get; set; } = string.Empty;
     public int TotalProducts { get; set; }
-    public int TotalQuantity { get; set; }
+    public decimal TotalQuantity { get; set; }
     public int LowStockCount { get; set; }
     public decimal TotalValue { get; set; } // Quantity * AverageCost
     public List<BranchInventoryItemDto> Items { get; set; } = new();
@@ -20,7 +20,7 @@ public class BranchInventoryItemDto
     public string ProductName { get; set; } = string.Empty;
     public string? ProductSku { get; set; }
     public string? CategoryName { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public int ReorderLevel { get; set; }
     public bool IsLowStock { get; set; }
     public decimal? AverageCost { get; set; }
@@ -37,7 +37,7 @@ public class UnifiedInventoryReportDto
     public string ProductName { get; set; } = string.Empty;
     public string? ProductSku { get; set; }
     public string? CategoryName { get; set; }
-    public int TotalQuantity { get; set; }
+    public decimal TotalQuantity { get; set; }
     public decimal? AverageCost { get; set; }
     public decimal? TotalValue { get; set; }
     public int BranchCount { get; set; }
@@ -49,7 +49,7 @@ public class BranchStockDto
 {
     public int BranchId { get; set; }
     public string BranchName { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public int ReorderLevel { get; set; }
     public bool IsLowStock { get; set; }
 }
@@ -65,7 +65,7 @@ public class TransferHistoryReportDto
     public int CompletedTransfers { get; set; }
     public int PendingTransfers { get; set; }
     public int CancelledTransfers { get; set; }
-    public int TotalQuantityTransferred { get; set; }
+    public decimal TotalQuantityTransferred { get; set; }
     public List<TransferSummaryDto> Transfers { get; set; } = new();
     public List<BranchTransferStatsDto> BranchStats { get; set; } = new();
 }
@@ -78,7 +78,7 @@ public class TransferSummaryDto
     public string FromBranchName { get; set; } = string.Empty;
     public string ToBranchName { get; set; } = string.Empty;
     public string ProductName { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string Status { get; set; } = string.Empty;
     public string Reason { get; set; } = string.Empty;
     public DateTime? CompletedAt { get; set; }
@@ -90,9 +90,9 @@ public class BranchTransferStatsDto
     public string BranchName { get; set; } = string.Empty;
     public int TransfersSent { get; set; }
     public int TransfersReceived { get; set; }
-    public int QuantitySent { get; set; }
-    public int QuantityReceived { get; set; }
-    public int NetChange { get; set; } // QuantityReceived - QuantitySent
+    public decimal QuantitySent { get; set; }
+    public decimal QuantityReceived { get; set; }
+    public decimal NetChange { get; set; } // QuantityReceived - QuantitySent
 }
 
 /// <summary>
@@ -114,9 +114,9 @@ public class LowStockItemDto
     public string ProductName { get; set; } = string.Empty;
     public string? ProductSku { get; set; }
     public string? CategoryName { get; set; }
-    public int TotalQuantity { get; set; }
+    public decimal TotalQuantity { get; set; }
     public int TotalReorderLevel { get; set; }
-    public int Shortage { get; set; }
+    public decimal Shortage { get; set; }
     public decimal? AverageCost { get; set; }
     public decimal? EstimatedRestockCost { get; set; }
     public List<BranchLowStockDetailDto> BranchDetails { get; set; } = new();
@@ -126,9 +126,9 @@ public class BranchLowStockDetailDto
 {
     public int BranchId { get; set; }
     public string BranchName { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public int ReorderLevel { get; set; }
-    public int Shortage { get; set; }
+    public decimal Shortage { get; set; }
     public bool IsCritical { get; set; } // Quantity = 0
 }
