+++ backend/KasserPro.Application/DTOs/Reports/ReportDto.cs
@@ -67,7 +67,7 @@ public class TopProductDto
 {
     public int ProductId { get; set; }
     public string ProductName { get; set; } = string.Empty;
-    public int QuantitySold { get; set; }
+    public decimal QuantitySold { get; set; }
     public decimal TotalSales { get; set; }
 }
